refactor(auto_reconnector): remove invalid servers from reconnect
This patch introduces the concept of a `join_result_t` to intra
cluster connection attempts. This enum is used (for now) during
direct calls to `join_blocking`, in order to make logical decisions
based on the join result. Primarily this information is used by
the `auto_reconnector_t` in order to prevent connection attempts to
servers that have known, unrecoverable errors during the handshake